Extract WpHTMLUtil
